+++ libft/src/lists/__set.c
@@ -6,7 +6,7 @@
 /*   By: tlouro-c <tlouro-c@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/27 16:37:24 by tlouro-c          #+#    #+#             */
-/*   Updated: 2024/01/08 00:00:28 by tlouro-c         ###   ########.fr       */
+/*   Updated: 2024/01/08 00:04:45 by tlouro-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
